+++ tests/testthat/test_print_Logger.R
@@ -4,17 +4,24 @@ context("format")
 
 test_that("formatting Loggers works as expected", {
 
+  tf1 <- tempfile()
+  tf2 <- tempfile()
+  tf_long <- tempfile(pattern = paste(letters, LETTERS, sep = "-", collapse = "-"))
+  on.exit(file.remove(tf1, tf2, tf_long))
+
   l <- Logger$new(
     "test_logger",
     appenders = c(
-      AppenderFile$new(file = tempfile()),
+      AppenderFile$new(file = tf1),
       AppenderConsole$new(),
-      AppenderFile$new(threshold = 100, file = paste0(tempfile(), tempfile(), tempfile())),
+      AppenderFile$new(threshold = 100, file = tf_long),
       AppenderDt$new(),
       AppenderBuffer$new(
-        appenders = list(AppenderDt$new(), AppenderFile$new(tempfile())))
+        appenders = list(AppenderDt$new(), AppenderFile$new(file = tf2))
+      )
     )
   )
+
 
   # ensure that print doesn't raise exceptions
   expect_output(print(l))
@@ -22,7 +29,3 @@ test_that("formatting Loggers works as expected", {
   expect_output(print(Logger$new("blubb", propagate = FALSE)))
   expect_output(print(Logger$new("blubb", propagate = FALSE, appenders = Appender$new())))
 })
-
-
-
-